Update login styling and merge template files
